Complete Post

+++ homepage/src/main/java/com/example/homepage/PostDir/posttag/repository/PostTagRepository.java
@@ -3,5 +3,5 @@ package com.example.homepage.PostDir.posttag.repository;
 import com.example.homepage.PostDir.posttag.entity.PostTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostTagRepository extends JpaRepository<PostTag, Integer> {
+public interface PostTagRepository extends JpaRepository<PostTag, Long> {
 }
