+++ homepage/src/main/java/com/example/homepage/RecruitDir/recruit/repository/RecruitRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RecruitRepository extends JpaRepository<Recruit, Integer> {
+public interface RecruitRepository extends JpaRepository<Recruit, Long> {
     Optional<Recruit> findFirstByOrderByIdAsc();
 }
